Fix onnx warnings in verify and parse tests

+++ test/onnx/gen_onnx.py
@@ -11550,7 +11550,7 @@ def softmaxcrossentropyloss_2d_mean_reduction_half_test():
 def softmaxcrossentropyloss_2d_no_reduction_weighted_test():
     scores = helper.make_tensor_value_info('0', TensorProto.FLOAT, [4, 4])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4])
-    weights = helper.make_tensor_value_info('2', TensorProto.DOUBLE, [4])
+    weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
     loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
@@ -11613,7 +11613,7 @@ def softmaxcrossentropyloss_2d_no_reduction_half_weighted_test():
 def softmaxcrossentropyloss_2d_sum_reduction_weighted_test():
     scores = helper.make_tensor_value_info('0', TensorProto.FLOAT, [4, 4])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4])
-    weights = helper.make_tensor_value_info('2', TensorProto.DOUBLE, [4])
+    weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
     loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
@@ -11676,7 +11676,7 @@ def softmaxcrossentropyloss_2d_sum_reduction_half_weighted_test():
 def softmaxcrossentropyloss_2d_mean_reduction_weighted_test():
     scores = helper.make_tensor_value_info('0', TensorProto.FLOAT, [4, 4])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4])
-    weights = helper.make_tensor_value_info('2', TensorProto.DOUBLE, [4])
+    weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
     loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
@@ -11785,7 +11785,7 @@ def softmaxcrossentropyloss_kd_no_reduction_weighted_test():
                                            [4, 4, 2, 2])
     labels = helper.make_tensor_value_info('1', TensorProto.INT32, [4, 2, 2])
     weights = helper.make_tensor_value_info('2', TensorProto.FLOAT, [4])
-    loss = helper.make_tensor_value_info('3', TensorProto.DOUBLE, [4])
+    loss = helper.make_tensor_value_info('3', TensorProto.FLOAT, [4])
 
     node = onnx.helper.make_node(
         "SoftmaxCrossEntropyLoss",
